// types/socket.ts
export enum CHAT_SOCKET_ROOM {
  USERS = "USERS",
  CHATS = "CHATS",
}

export enum CHAT_SOCKET_EVENT {
  NEW_MESSAGE = "NEW_MESSAGE",
  UPDATE_MESSAGE = "UPDATE_MESSAGE",
  NEW_REACTION = "NEW_REACTION",
  NEW_REPLY = "NEW_REPLY",
  MESSAGE_DELETED = "MESSAGE_DELETED",
  USER_SEEN = "USER_SEEN",
  MESSAGE_SEEN = "MESSAGE_SEEN",
  MESSAGE_RECEIVED = "MESSAGE_RECEIVED",

  USER_TYPING = "USER_TYPING",
  USER_JOINED = "USER_JOINED",
  USER_LEFT = "USER_LEFT",
  NEW_CHAT = "NEW_CHAT",
}

export enum CONNECTION_SOCKET_EVENT {
  NEW_CONNECTION = "NEW_CONNECTION",
  CONNECTION_ACCEPTED = "CONNECTION_ACCEPTED",
  CONNECTION_CANCELED = "CONNECTION_CANCELED",
}

export enum NOTIFICATION_SOCKET_EVENT {
  NEW_NOTIFICATION = "NEW_NOTIFICATION",
  NOTIFICATION_READ = "NOTIFICATION_READ",
  UNREAD_COUNT_UPDATE = "UNREAD_COUNT_UPDATE",
}